Add cmake support. Minor cleanups to dispatch, set "has_idle" correctly.
git-svn-id: https://protobuf-c.googlecode.com/svn/trunk@276 00440858-1255-0410-a3e6-75ea37f81c3a

+++ src/google/protobuf-c/protobuf-c-rpc.c
@@ -31,8 +31,8 @@
 /* enabled for efficiency, can be useful to disable for debugging */
 #define RECYCLE_REQUESTS                1
 
-#define UINT_TO_POINTER(ui)     ((void*)(ui))
-#define POINTER_TO_UINT(ptr)     ((unsigned)(ptr))
+#define UINT_TO_POINTER(ui)      ((void*)(uintptr_t)(ui))
+#define POINTER_TO_UINT(ptr)     ((unsigned)(uintptr_t)(ptr))
 
 #define MAX_FAILED_MSG_LENGTH   512
